fix: only init if not already

+++ init.sh
@@ -1,13 +1,20 @@
 #!/usr/bin/env bash
 
-Rscript -e 'renv::init(bare = FALSE)'
-Rscript -e 'renv::install("rmarkdown")'
-Rscript -e 'renv::snapshot(type = "all")'
 
-python3 -m venv .venv
-source .venv/bin/activate
-python3 -m pip install jupyter
-python3 -m pip freeze > requirements.txt
+if [ ! -f "renv.lock" ]; then
+  Rscript -e 'renv::init(bare = FALSE)'
+  Rscript -e 'renv::install("rmarkdown")'
+  Rscript -e 'renv::snapshot(type = "all")'
+fi
 
-julia -e 'using Pkg; Pkg.activate("."); Pkg.instantiate()'
-julia --project=. -e 'using Pkg; Pkg.add("IJulia")'
+if [ ! -f "requirements.txt" ]; then
+  python3 -m venv .venv
+  source .venv/bin/activate
+  python3 -m pip install jupyter
+  python3 -m pip freeze > requirements.txt
+fi
+
+if [ ! -f "Project.toml" ]; then
+  julia -e 'using Pkg; Pkg.activate("."); Pkg.instantiate()'
+  julia --project=. -e 'using Pkg; Pkg.add("IJulia")'
+fi
